Optionally apply the aberration correction at full accuracy.
The default behavior (fast_aber=True) uses a small-angle approximation in
constructing the aberration rotation quaternion.  This approximation is valid
for typical values of the annual and diurnal aberration corrections, as these
are fractions of a degree in scale.

When fast_aber=False, the aberration quaternion is computed using the exact
quaternion rotation function.